+++ cache_headers/middleware.py
@@ -4,7 +4,9 @@ import re
 import uuid
 
 from django.conf import settings
+from django.contrib.auth.signals import user_logged_in, user_logged_out
 from django.core.cache import cache
+from django.core.exceptions import SuspiciousOperation
 from django.http import HttpResponseRedirect
 
 from cache_headers import policies
@@ -37,6 +39,13 @@ for cache_type in TIMEOUTS.keys():
 # Sort from longest string to shortest
 rules.sort(key=lambda x: x[3], reverse=True)
 
+# Subscribe to signals so we can mark the request
+def on_user_auth_event(sender, user, request, **kwargs):
+    setattr(request, "_dch_auth_event", True)
+
+user_logged_in.connect(on_user_auth_event)
+user_logged_out.connect(on_user_auth_event)
+
 
 class CacheHeadersMiddleware(object):
     """Put this middleware before authentication middleware because response
@@ -48,44 +57,46 @@ class CacheHeadersMiddleware(object):
         if settings.DEBUG:
             return response
 
-        # Record whether cache control is set
-        has_cache_control = False
+        # If cache control was set at the start of this method then do nothing
         if ("Cache-Control" in response) or ("cache-control" in response):
-            has_cache_control = True
+            return response
 
         # Do nothing if response code is not 200
         if response.status_code != 200:
             return response
 
-        # Default
+        # Default policy is to not cache
         response["Cache-Control"] = "no-cache"
 
+        # If there is no user on the request then do nothing
         user = getattr(request, "user", None)
         if not user:
-            response["Cache-Control"] = "no-cache"
             return response
 
-        # During login we need to be careful to not accidentally cache. The
-        # isauthenticated cookie helps to determine the state.
-        if user.is_authenticated() and not request.COOKIES.get("isauthenticated"):
-            expires = request.session.get_expiry_date()
-            response.set_cookie("isauthenticated", 1, expires=expires)
-            response["Cache-Control"] = "no-cache"
+        # During login and logout we do not cache
+        if hasattr(request, "_dch_auth_event"):
             return response
 
-        if not user.is_authenticated() and request.COOKIES.get("isauthenticated"):
-            response.delete_cookie("isauthenticated")
-            response["Cache-Control"] = "no-cache"
-            return response
+        # If user is anonymous but SESSION_COOKIE_NAME is in cookies and has a
+        # value then this is an attempt at tampering.
+        if user.is_anonymous() and (settings.SESSION_COOKIE_NAME in request.COOKIES):
+            cookie = request.COOKIES[settings.SESSION_COOKIE_NAME]
+            if getattr(cookie, "value", cookie):
+                raise SuspiciousOperation(
+                    "User is anonymous but received a sessionid"
+                )
 
         # Never cache non-GET
         if request.method.lower() not in ("get", "head"):
             return response
 
-        # If cache control was set at the start of this method then do nothing.
-        # We only return here because isauthenticated needs opportunity to be
-        # set.
-        if has_cache_control:
+        # Don't cache if response sets cookies
+        if response.has_header("Set-Cookie"):
+            logger = logging.getLogger("django")
+            logger.warn(
+                "Attempting to cache path %s but Set-Cookie is on the response" \
+                    % request.get_full_path()
+            )
             return response
 
         # Determine age and policy. Use cached lookups.
@@ -130,14 +141,6 @@ class CacheHeadersMiddleware(object):
         if age:
             policy = POLICIES[cache_type]
             policy(request, response, user, age)
-
-            # Warn on potentially erroneous usage
-            if "Set-Cookie" in response:
-                logger = logging.getLogger("django")
-                logger.warn(
-                    "Caching path %s but Set-Cookie is on the response" \
-                        % full_path
-                )
         else:
             response["Cache-Control"] = "no-cache"